+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   linked object file in which the given symbol is defined.
 - `msgpack::ArrayWriter` helper struct for generating msgpack arrays from
     arbitrary serializable data.
+- `msgpack::ValueIter` helper struct for iterating over msgpack values.
 - `tarantool::network::client` alternative async network client.
 - `tarantool::network::client::reconnect` reconnecting async network client based on `network::client`.
 - `tarantool::network::protocol` sans-io (without transport layer) implementation of Tarantool Binary Protocol.